+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # changelog
 
 * :warning: **BREAKING CHANGES**: remove deprecated functions in `basicalter`: `UniqueStrings`, `DelStringInSlice`, `FilterStringsWith`, `ReverseStrings`, `ReplaceStringsWith`
+* add new `basicnew` package with this functions:
+  * `MapKeys`: MapKeys generate a new slice with all keys of a map.
+  * `MapValues`: MapValues generate a new slice with all values of a map.
 
 ## v0.6.0